+++ packages/zowe-explorer-api/src/profiles/ProfilesCache.ts
@@ -185,7 +185,15 @@ export class ProfilesCache {
     ): void {
         // Note: When autoStore is disabled, nested profiles within this service profile may not have their credentials updated.
         const profIndex = this.allProfiles.findIndex((profile) => profile.type === profileLoaded.type && profile.name === profileLoaded.name);
+
+        // Warn users that the profile is not cached
+        if (profIndex === -1) {
+            this.log.warn(`Profile ${profileLoaded.name} of type ${profileLoaded.type} is not cached.`);
+            return;
+        }
+
         this.allProfiles[profIndex].profile = profileLoaded.profile;
+
         const defaultProf = this.defaultProfileByType.get(profileLoaded.type);
         if (defaultProf != null && defaultProf.name === profileLoaded.name) {
             this.defaultProfileByType.set(profileLoaded.type, profileLoaded);